// src/app/core/models/brand.model.ts
export interface Brand {
  Make_ID: number;
  Make_Name: string;
}
export interface VehicleModel {
  Make_ID: number;
  Make_Name: string;
  Model_ID?: number;
  Model_Name: string;
}
export interface VehicleType {
  MakeID: number;
  MakeName: string;
  VehicleTypeId: number;
  VehicleTypeName: string;
}